+++ src/main/java/co/invest72/investment/console/input/parser/InstallmentInvestmentAmountParser.java
@@ -1,5 +1,7 @@
 package co.invest72.investment.console.input.parser;
 
+import java.math.BigDecimal;
+
 import co.invest72.investment.domain.InvestmentAmount;
 import co.invest72.investment.domain.amount.AmountType;
 import co.invest72.investment.domain.amount.MonthlyInstallmentInvestmentAmount;
@@ -19,7 +21,7 @@ public class InstallmentInvestmentAmountParser implements InvestmentAmountParser
 			throw new IllegalArgumentException("투자 기간 단위는 'MONTHLY' 또는 'YEARLY'이어야 합니다.");
 		} else if (periodType.equals(AmountType.MONTHLY.getDescription())) {
 			int amount = Integer.parseInt(parts[1]);
-			return new MonthlyInstallmentInvestmentAmount(amount);
+			return new MonthlyInstallmentInvestmentAmount(BigDecimal.valueOf(amount));
 		} else {
 			int amount = Integer.parseInt(parts[1]);
 			return new YearlyInstallmentInvestmentAmount(amount);
